More type information and better docs
I've now propagated docstrings and descriptions
to OpenAPI and Thing Description. I've also
now described types (in/out and property) in
OpenAPI and Thing Description.

Some edge cases remain unsupported, in
particular multiple types in an array, or
the ability to specify whether extra properties
are allowed or not in the TD.

I suspect this has little impact - the OpenAPI
docs will be correct, and the TD will either be
OK, or will raise an error if something is badly wrong.