Worked around a possible pydantic bug
I was getting strange errors where required fields showed up as
required=False, default=Ellipsis. This seems to be a bug in
pydantic.create_model (or at least a confusion in my use of it).
I have reported it as https://github.com/pydantic/pydantic/issues/7064.
For now, replacing `...` with `Field(...)` when creating required
fields seems to have fixed the problem.