Added support for file output and thing lifecycle
Things that are context managers will now
be entered and exited by the app. This
provides a neat, testable way to make sure
hardware resources are opened and closed
cleanly.

Actions may now output files using FileManager, which will take care of
storing and returning them.

NB FileManager does not currently dispose
of files, which will be a problem in production!