Sorted out lifecycle and concurrency
Things can now be context managers, which will correctly
handle start-up and shut-down code.
I've simplified logic around property change events: we
now *always* use a blocking portal, and it's an error to generate events
without one.
In the future, I may relax this - I think it's fair enough
to only propagate events while an event loop is running,
but perhaps it's reasonable to allow setting properties
in __init__ (when there may not be an event loop).