WIP: Python client
Currently this generates an object with a function for
each action and a property for each property.

It does not yet use forms properly, so URLs are
only right by accident.

It also doesn't set action parameters in the function signature
yet - or type the properties.

The missing link is a DataSchema to type hint
converter.