Dependency injection!
This is a mega-commit, but it adds FastAPI dependency
injection to actions. This replaces the ad-hoc mechanism
I was using for FileManager, and should make it possible
for Things to depend on other Things in the same server.

Inter-thing dependency is implemented (for now) with a
special client class - so actions won't require you to supply
dependencies, just what would be the payload.
The internal client class doesn't currently handle injecting
dependencies, so it will fail if a function needs e.g. a
FileManager, this should in principle be solvable in the
future with one or two caveats (specifically, we won't
propagate dependencies of client Things to the signature
of action functions - but if the relevant dependencies exist
already (or can be constructed from the Request object), it
ought to be possible to supply them.