Support for blob outputs from actions
Actions may now declare a return type derived from
BlobOutput. This will serialise to a JSON link specifying the
media type, and will allow us to return a raw (can be binary)
response.

In the future, we should use the `accept` HTTP header to pick
whether to return JSON or the specified mime type. We
should also implement behaviour that makes the
200 status code reachable from an action invocation.