Comment out unnecessary locking error
I added code to raise an exception if we overwrite a
frame in the MJPEG circular buffer that's being read.
As we are simply replacing one bytes object with another,
there's no actual memory conflict going on - and the
exceptions were crashing the Pi Camera.

It would be good to either properly remove this, or fix
it so that we don't end up overwriting frames before they
are handed back by the Response object during garbage
collection. This probably requires a better understanding
of what happens when a client stops reading the frames.