Remove ringbuffer locking code
The ringbuffer in an MJPEGStream uses `bytes` objects and
replaces them each time a new frame is added. This means
that we don't need to worry about replacing a buffer that
is being read, as the copy will continue to exist until it's no
longer needed.

The exceptions this raised were causing a crash in the
picamera2 background thread, which was difficult to
handle properly: I think it's best to remove this code, as
it's unnecessary, confusing, and problematic.

I will leave the context manager for accessing frames, in
case it becomes helpful again in the future.