Switch to pydantic_numpy
pydantic_numpy now does everything I need it to. There's a slight difference in how we have handled 0d arrays, but I think
this is an edge case I can ignore.

I've defined, for now, an NDArray type alias that is a work-alike
for my old class. However, this should be phased out in favour
of `pydantic_numpy.typing` which is much better.