Switch back from pydantic-numpy
Switching to pydantic-numpy broke Thing Description
generation, as the generated JSON schema was not valid.
I've switched back to my horribly clunky but just-about-working
NDArray type.

A useful to-do would be to add specific types for 1D ... 6D arrays. These types would account for the majority of use cases, and could even be generic (to handle int/float/whatever).

I'd still like to support pydantic-numpy in the future, but will
need to do some thinking about how to generate sensible
data schemas for the Thing Description.