Allow blobs as input
This commit builds on the last few, and enables BlobOutput objects to be passed around the server. This should enable
**both client and server code** to do things like:
blob = thing.action1()
result = thing.action2(image=blob)

There are unit tests for both client-side and server-side chaining.

Currently there is one serialiser that's a stub, and we need to check for that, but currently there's no unit test doing so.

Being able to pass BlobOutputs around more freely will be very useful for handling images, for example.