Fix NDArray Type Annotation
After much confusion around TypeAlias types, this is fixed by adding
return_type to the PlainSerialiser annotation.

Pydantic 2.10 is stricter about forward references in type annotations. This
becomes a problem for NDArray, because the NestedListOfNumbers
type it returns isn't present in the
scope of the model (which probably
only imported NDArray).

The solution to this is to specify return_type
explicitly, rather than infer it based on the
annotation of the serialiser function.

Using the `type` keyword might help in the future, but is currently only Python 3.12 and up.

As part of this I realised there was some incorrect logic in
`type_to_dataschema` which was
using TypeAdapter on BaseModel
subclasses, which is ineefficient.

A test is added that should fail if this issue resurfaces.