Enforce type hints
This commit changes ThingProperty to raise an error
if it is used without type hints. I have updated the test code accordingly, including tests
for un-annotated properties raising an error.

I've also tweaked the type hints on the @thing_property and @thing_setting decorators. MyPy was giving errors that their
argument was a subtype of Thing rather than Thing, so I have now used `Callable[..., Value]`
(i.e. the arguments are not checked, but the return value is).