Remove the file_manager module
`file_manager` was a previous attempt to handle actions that created files. It is retired in favour of the `Blob` class. I'm not aware of any examples
of `file_manager` being used "in the wild" and it
does not do anything that `Blob` can't do.

We are therefore getting rid of it as vestigial code.