Fix a couple of typing issues I introduced
I thought adding a type annotation to `Invocation.action` would be a simple improvement, but it confuses mypy badly.

I have removed the annotation and the error is gone, but we should fix it properly later - I've added a note to NOTES.md.

I also forgot to delete the `as_responses` argument from a call to `invocations()`, this is now fixed. Thanks for spotting that, mypy.