Delete examples
We may want to remove the folder entirely. The counter is moved into the tutorial, and the demo thing server was horrible anyway.
The `MyThing` class exists in the module as an example thing, but should probably be split up and most of it moved into unit tests.

In short, this deletion doesn't remove anything that isn't provided in a more useful form elsewhere.