lock dependencies for py3.10 and exclude old python from flake8
dev-dependencies.txt was regenerated under python 3.10, so numpy is no longer 2.3 and will install.

I've set flake8 to only run under python 3.12 and 3.13 - it fails on older Python, but passes with no errors on 3.12 and 3.13.
Given that we're linting rather than testing, I think there's little advantage to doing it four times, so I will disable that step on 3.10 and 3.11, and spend the time fixing something else.