Build docs on 3.12
The Sphinx error is the same as the flake8 error on Python 3.11. Changing the build version on readthedocs should fix the problem.

I also now need to install the package for readthedocs (as I import it in `conf.py` to exclude some symbols). I've added
`python.install.path: .` which I hope will do that.