First pass at splitting up ThingProperty
This commit swaps ThingProperty for two descriptor classes, FunctionalProperty (using getters/setters) and
DataProperty (variable-like). This preserves the two
useful behaviours out of the six (!) possible behaviours
implemented by the old ThingProperty.

I've also added a `property` function that provides a single interface to both: it may be used either as a
decorator, or as a field specifier.

This does not yet pass tests/type checking.