Add typing tests for thing definition code.
Currently, mypy checks the package for type safety. It would be really useful to check that code using the library will be typed
correctly, so that Thing subclasses will play nicely with mypy and friends.

This commit starts that process, by adding a folder of test code that will be run through mypy, with `--warn-unused-ignores`.

Activating the `unused-ignores` error means that we can check that code raises an error just by adding a # type: ignore
comment. If the error isn't raised, the unused # type: ignore will cause an error.

This is not yet passing, because the getter/setter form of property triggers an error, which seems to be a difficult one to
work around. This has been described previously:
https://github.com/python/typing/discussions/1102