Fixed flake8 errors
I've globally disabled some whitespace-checking as this is done better by ruff (in particular, it handles @overload better).

I've disabled a few more warnings related to @overload definitions line-by-line, with a block comment explaining why at the top of the module. This is mostly flake8 not understanding @overload or @builtins.property.