Fix test of BaseDescriptor and docs
`test_docs.py` was failing because of an incompatibility between `WindowsPath` and `inspect.getsource`. I've fixed this by converting the path to a string, and also handled the exception
in get_class_attribute_docstrings (by returning an empty dict). I think this is reasonable: it's not an error that can be recovered from, and there is never any guarantee that docstrings are not `None`.

The BaseDescriptor test was failing simply because I defined a mock function incorrectly - now fixed.