Enable Bugbear linter rules and fix errors
I've applied lots of minor fixes that shouldn't cause any changes in code behaviour. The only fixes that are likely to
affect anything were in the previous commit, related to
turning a few things into dataclasses.

I have not enforced B008, because it is incompatible with
FastAPI Depends() (see comment in pyproject.toml).
There's a work-around to use a plugin for flake8 that exempts
Depends() specifically, but that won't work for Ruff.