Fix a race condition in test_actions_list
test_actions_list sometimes failed, because the invocation expired and was deleted before we called `/action_invocations/` to list the invocations.

I believe the problem was that I used an action with a very short retention time (so I could test expiration of the invocation), and it may have expired before I listed the actions.
I've now added an action with a longer retention time. If a full second is elapsing between the two requests, something else has gone quite wrong.