Don't raise AttributeError, only KeyError
I've added a default to the `getattr` call, so that we don't raise
`AttributeError` if we attempt to observe something that's
not an attribute. Instead, we will raise KeyError, which is
the same as if we attempted to observe something that's
not a property.

I think it makes sense not to distinguish between missing
entirely and not-a-property: the client shouldn't be aware
if something's a Python attribute that isn't exposed
as an affordance.

I raise KeyError because I'm thinking the string passed in
is being used as a key to look up properties.