Rework and improve tests and documentation
I've gotten rid of the previous tests in favour of code that
tests the same things much more cleanly. Changes include:

1. Not using MyThing, instead using a minimal Thing
    defined in the module. This might be moved to a
    common things_to_test module at some point.
2. Using fixtures better, in particular making use of yield
    to ensure everything's cleaned up properly after each test,
    and to avoid the need for tests to call functions that do
    parts of the test (rather than just the set-up).
3. Docstrings to explain what all the tests are doing.

I also found and fixed a bug, where the websocket test for
observing an action hung indefinitely, as we were waiting for
a message that never came.

Frustratingly, there is no timeout option for Starlette test
websocket connections.