Exempt sphinx config from annotations rules.
Sphinx is configured with a Python script, but this is not part
of the package, the examples, or the test code. I've therefore
excluded it from the rules.

I've also changed the exclusion to be specific to the docs config file, and not the whole docs folder, just in case we
add scripts there in the future.

This also means any example Python files will be checked.