Don't catch blind Exceptions
I've enabled a Ruff rule to disallow `except Exception:` and
removed one instance of that from the codebase.

I have left one `except Exception:` in test code (which captures
exceptions in a thread for later analysis), and another in the
MJPEG stream (which logs exceptions but does not crash the
server if they happen in the MJPEG stream).

The MJPEG code might be able to be eliminated now that
streams stop more gracefully - but I don't want to mess with
that until we're in a position to test it more extensively.
I've swapped the `logging.error` for `logging.exception` so
that the full stack trace is dumped to the log.