Convert an assert to an exception
This was guarding a weakref to the Thing: this isn't an error we
expect, or an error we should handle, so I've excluded the
if block from testing. Coverage is unchanged: the assert
statement was equivalent to the new block, it just didn't
show as uncovered if the asertion passed.