Fix failing tests of the server
Exceptions raised when setting up things were being wrapped in an ExceptionGroup: this is fixed by using pytest.RasisesGroup, but that required updating pytest.

The test that checked for errors when a non-Thing was added to the server was failing, because the config was invalid.
I've fixed it, but we really ought to use a model here.

I've also renamed the import to avoid confusion with
`server`, so we now import the server module as `ts`.