Change `add_thing` to use name and instantiate Things.
In order to improve the Thing lifecycle, we're going to move
to having the server instantiate the Thing. This will allow
us to inject arguments to `__init__` that simplify the set-up
process.

This commit changes `add_thing` to accept a class rather than an object, and adds `args` and `kwargs` parameters to enable
values to be passed to `__init__`.

This commit swaps the `path` for a `name` (in all known cases this simply removes the leading `/` from the string) in preparation for making a distinction between names and paths.

This commit does not change `Thing` at all: that will come later. It also doesn't change the test code - that will touch many more files!