Update test suite to use new add_thing syntax.
I've taken this opportunity to move nearly all of the server
set-up into fixtures. It would be nice to deduplicate this
code in the future, but I think it's already neater than it was.

Some tests use things that are not connected to a server.
These will require mocking in the future.