Implement ThingServerInterface
Each Thing is now provided with a ThingServerInterface which allows it to interact with the server. The ThingServerInterface is
specific to each Thing, and is how the Thing obtains its name and path.

This removes the need to pass around the BlockingPortal: the
ThingServerInterface provides a method to call async code in
the event loop on the server. I hope this is clearer.

The ThingServerInterface takes care of making sure we're connected to a server: I've therefore been able to remove a lot
of error-handling code and tests for NotConnectedToServerError.
This will need to be tested for explicitly, on the ThingServerInterface.