Update test suite to mock the server interface.
Now that we require a ThingServerInterface to create a Thing,
I have provided a convenience test harness,
`create_thing_without_server()` to consistently and cleanly allow
`Thing` subclasses to be instantiated in test code.

This auto-generates a name and path for the Thing, and mocks the required functionality.

I have removed some tests that were checking for conditions where the Thing was not connected. Things are now created
in a connected state, so this isn't needed any more.

I should add tests for ThingServerInterface, particularly to make sure that it raises sensible errors when the event loop
is not running. That's for a future commit.