Mark slow tests to speed up repeated testing.
Some tests (specifically those that spin up an HTTP server) are
much slower than the rest. I have marked these as "slow" so
the test suite may be run with `-m "not slow"` when repeatedly
re-testing.

Obviously we should run all the tests in CI, this is a convenience
feature for development.