Create invocation loggers before starting actions.
This came to light during testing of the new ThingServerInterface..
There's no reason errors setting up the logger should
be swallowed by a catch-all exception: errors there are
a bug in LabThings.

Also, setting up the logger after emitting an event meant
that an exception raised there caused another exception
because of the missing logger.

I have moved the creation of the logger out of the `try:` block
to fix both these issues.