Use ThingServerInterface in settings code
This change makes it possible to load settings during __init__,
though it does not yet do so. As the settings path is always
available, we can remove some error-checking code that
handled the period between the Thing being created and
it having a settings path to use: this period no longer exists.