Deduplicate `path_for_thing`.
I had the same logic in `ThingServer.path_for_thing` and also `ThingServerInterface.path`. I've now called the former from the latter.

I've then duplicated the logic again in `MockThingServerInterface` but I am less concerned about that, as duplication in a mock is not such a problem.

There is a test to check the mock server interface generates a sensible path, and a test that asking for the path of a Thing
that's not on the server raises an error.