Test code for thing_connection
This checks that two things can be connected to each other.
In writing this test, I realised my initial assumption that it
would be easy to have circular dependencies and that these
would not cause a problem wasn't true.

See next commit for my solution to the circular dependency
problem.

The test covers the expected functionailty and anticipated errors in what I hope is a fairly bottom-up way.