Testing and improvements
I've covered 100% of `thing_connections.py` with unit
tests, with bottom-up tests that don't use the server
or `Thing`, as well as more realistic tests that do use `Thing` and `ThingServer`.

This uncovered a few issues that I've fixed, in particular with `thing_type` and with error handling
logic in `connect`.

I've now split `connect` into `_pick_things` (which finds the Thing(s) to connect) and `connect` (which makes sure they match the type hint and supplies context for error messages). I think this is clearer.

I think the split also makes it more testable, as I can test the logic in `_pick_things` separately.