Typing fixes.
I had hoped not to need to ignore these. It would probably be possible if I made ThingConnection
non-generic, but I think it is worth keeping that
feature: it allows us to specify a subscripted
class, if for whatever reason it's better to do that
than use a type hint on the attribute.

I have added explanation in the docstring as to why
I have ignored type checking, and why I think it's
the right thing to do.