Explicitly don't support forward references in type subscripts.
typing.get_type_hints automatically resolves forward references
using sensible values for locals and globals.
This is used to permit forward references in field-typed descriptors.

There's no corresponding way to resolve forward references in type subscripts, so for now I
have made this an error - descriptors that are subscripted may
not be subscripted with a string.

Given that I don't anticipate type subscripts will be used much, this
is probably not a major limitation.
It could be fixed by implementing our own type evaluator
based on `get_type_hints` but I would prefer not to do this if I can avoid it.