Don't make loggers per-invocation
This commit adds a logger for every Thing instance, and a
custom log handler and filter that inject the invocation ID
into LogRecord objects.

This means we can still filter out invocation logs as we did before, but we no longer need to make
a new logger for each invocation.

This more or less follows the example given in the Logging Cookbook for adding context.