Initialise all Things during ThingServer.__init__
This marks a few more methods of ThingServer as private, and accepts a dict of Things to be created during initialisation.

In order to do this cleanly, I have now formalised a schema for config files, using a Pydantic model.

Tests (and no doubt fixes) will come in the next commit.