Tidy up config models and make them available at module level.
The config modules now do all the required validation, including Thing names,
ensuring we expand classes into full `ThingConfig` objects, and supplying
default values for args, kwargs, etc.

I've exposed ThingConfig and ThingServerConfig at module level as they are
likely useful in a number of scenarios, particularly writing tests.

It is possible to supply a dict whenever a ThingConfig is required - I'll include
this in the tests, but I don't want to recommend it - it's better to use the model
as it makes it harder to miss fields.