Update test code to pass Thing classes to ThingServer.__init__
I've removed ThingServer.add_thing in favour of passing a dictionary to ThingServer.__init__
and the test code is updated to reflect that.

As a rule, when thing instances were required, I would:
1. Create the server
2. Assign `my_thing = server.things["my_thing"]`
3. Assert `assert isinstance(my_thing, MyThing)`
That was enough for `pyright` (and I guess mypy) to correctly infer the type in my test code. It's slightly more verbose, but means we can use the same code in test and production, rather than
needing a different way to create/add things in test code.