Changes from review
With thanks to @julianstirling:

* Update copyright date in docs
* Mention URL of OpenAPI document and viewers
* Move the section on descriptors to developer notes.

I also spotted some broken links in documentation, which are fixed.

Additionally, moving the descriptor section out of "see also" left it a bit empty, so I've added some links to FastAPI and Pydantic. It's possible the "see also" section could be pruned instead.

I also spotted that we were missing a reference for settings (now added at the bottom of properties, pending its own page), and spotted a couple of out-of-date examples in `base_descriptor`.