Make blob_data_manager global
The BlobDataManager dependency (which is only ever used internally) used dependencies to obtain the ThingServer, and extracted the BlobDataManager from it.

This is now deprecated: Things can access the server via the ThingServerInterface.

The simplest fix is to make a global BlobDataManager object. This simply allows BlobData objects to be retrieved by UUID. In the future, it may become part of a `BlobData` base class, similar to `CancelEvent`.

This is only a change in behaviour if there is more than one `ThingServer` per Python instance. Even then, the only effect would be that blobs from one server could be accessed from the other server. That's not a supported scenario, and is very unlikely to lead to errors anyway - UUIDs should be unique, so there's no danger of blob IDs colliding.