Move decorator into descriptor module, and subclass BaseDescriptor
I've combined the decorator (`lt.fastapi_endpoint`) and the descriptor into one module.

I've also made the descriptor a subclass of `BaseDescriptor` to deduplicate handling of `name`, `title`, `description`. This does change behaviour slightly: we use the attribute name rather than the function name (the two are equal, when it's used as a decorator). Also, an attribute docstring may override the function docstring - though most likely there won't be an attribute docstring.