Refactor actions
This moves actions to being a single-file module, and moves the descriptor and decorator into it. This commit also makes a new module `invocations` that currently just holds the invocation model. It might make sense to move Invocation (the Thread subclass) in here to shrink Actions slightly.

I've refactored endpoints into a single module, which means the empty decorators and descriptors modules are now gone.