Get rid of vestigial code
We were creating a new subclass of ActionDescriptor each time we instantiated it in `lt.action`.
I'm fairly sure this was done because, in the past, we manipulated `__init__` or similar. That's no longer an issue.

I have added a test to make sure rewriting `self.__doc__` has the intended effect.