Fix type checking of actions
This introduces two changes:

1. actions are now properly type hinted.
2. ActionDescriptor now inherits from BaseDescriptor.

The significance of these two changes will be described at more length in the MR thread.

Inheriting from BaseDescriptor deduplicates some boilerplate, and deduplicates handling of name/title/description.

The main change is that we now use TypeVar and ParamSpec to fully describe the function being wrapped. ActionDescriptor is generic in three parameters, which allows us to pass the type of the bound function through the descriptor properly.

There was an assumption that the name of the function matches the name of the descriptor. This is now explicit.

Currently, the only fudge is that the type of the owning Thing isn't perfect. This is a limitation of BaseDescriptor, which doesn't have a generic parameter for the owning Thing.