Handle import errors more cleanly
This uses a WrapValidator to improve the handling of errors in an ImportString.

The problem, identified in #222, is that `TypeError` and `ValueError` are treated specially by `pydantic`.
This means that if either of those exceptions are raised during module import, we get a confusing validation error.

My wrap validator ensures that we get either a sensible import-related error (e.g. moule not found, name not found in module) or a single clear error stating that an error occurred during module import.

This does not dump a traceback, because doing so during validation tends to lead to very confusing messages.
However, it should make it obvious that the problem is an unimportable module, and point someone in the right direction.